# app/Http/Controllers/Api/OrderController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Http\Requests\CheckoutRequest;
use App\Http\Requests\UpdateOrderStatusRequest;
use App\Models\Order;
use App\Models\OrderItem;
use App\Http\Resources\OrderResource;
use App\Http\Resources\SellerOrderResource;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;


/**
 * @group Orders
 *
 * Endpoints for placing orders and managing order workflows.
 */
class OrderController extends Controller
{
    /**
     * List orders
     *
     * Returns orders for the authenticated user based on role visibility (admin, seller, or customer). Requires authentication.
     */
    public function index(Request $request){
        $user = $request->user();

        // Admin view
        if ($user->isAdmin()) {
            // Admin can see all orders, with relationships loaded
            return $this->success(OrderResource::collection(
                Order::with('user', 'items.product')->latest()->paginate(10)
            ));
        }

        // Seller view (orders that include at least one of their products)
        if ($user->isSeller()) {
            $orders = Order::with('items.product')
                ->whereHas('items', function ($query) use ($user) {
                    $query->whereHas('product', function ($query) use ($user) {
                        $query->where('user_id', $user->id);
                    });
                })
                ->latest()
                ->paginate(10);

                return $this->success(SellerOrderResource::collection($orders));
        }

        // Customer view
        $query = $user->orders()->with('items.product')->latest();
        if ($request->filter === 'active') {
            $query->whereIn('status', ['pending', 'processing', 'shipped']);
        } elseif ($request->filter === 'previous') {
            $query->whereIn('status', ['delivered', 'cancelled']);
        }
        return $this->success(OrderResource::collection($query->get()));
    }

    /**
     * Show order
     *
     * Returns details for a specific order if the authenticated user is allowed to view it. Requires authentication.
     */
    public function show(Request $request , Order $order){
        $order->load('items.product');

        $this->authorize('view', $order);

        return $this->success(new OrderResource($order));
    }

    /**
     * Update order status
     *
     * Updates the status of an order. Requires admin role.
     */
    public function updateStatus(UpdateOrderStatusRequest $request, Order $order){
        $this->authorize('update', $order);

        $validated = $request->validated();
        $order->update(['status' => $validated['status']]);

        return $this->success(new OrderResource($order), 'Order status updated successfully');
    }

    /**
     * Checkout
     *
     * Places a new order from the authenticated customer's cart and selected address. Requires customer role.
     */
    public function checkout(CheckoutRequest $request){
        $this->authorize('create', Order::class);

        $validated = $request->validated();
        $user = $request->user();

        // Ensure the address belongs to the user
        $address = $user->addresses()->where('id', $validated['address_id'])->first();
        if (!$address) {
            return $this->error('Invalid address', 400);
        }
        
        // Get the user's cart
        $cartItems = $user->cartItems()->with('product')->get();

        if ($cartItems->isEmpty()) {
            return $this->error('Cart is empty', 400);
        }

                        //---start transaction---
        return DB::transaction(function () use ($user, $cartItems, $address) {
            // Calculate total price
            $total = $cartItems->sum(function ($item) {
                return ($item->product->price * $item->quantity);
            });
            
            $fullAddress = "{$address->title},
                {$address->address},
                {$address->city}, 
                {$address->postal_code}, 
                {$address->country}";
            
            // Create Order Record
            $order = Order::create([
                'user_id' => $user->id,
                'status' => 'pending',
                'address' => $fullAddress,
                'total_price' => $total / 100,
            ]);
            //Move products from Cart to OrderItems
            foreach ($cartItems as $cartItem) {
                // Check stock one last time
                if ($cartItem->product->stock < $cartItem->quantity) {
                    abort(400, "Product {$cartItem->product->name} is out of stock");
                }
                // deduct stock
                $cartItem->product->decrement('stock', $cartItem->quantity);
                // Create OrderItem
                OrderItem::create([
                    'order_id' => $order->id,
                    'product_id' => $cartItem->product_id,
                    'quantity' => $cartItem->quantity,
                    'price' => $cartItem->product->price,
                ]);
            }
                // Clear user's cart
                $user->cartItems()->delete();
                
                return $this->success(['order_id' => $order->id], 'Order placed successfully', 201);
                //---end transaction---
        });
    }
}