# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'E-Commerce API'
  description: 'A full-featured e-commerce REST API built with Laravel 12.'
  version: 1.0.0
servers:
  -
    url: 'http://e-commerce-api.test'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can get your token by calling the login endpoint.'
security:
  -
    default: []
paths:
...
  /api/checkout:
    post:
      summary: Checkout
      operationId: checkout
      description: "Places a new order from the authenticated customer's cart and selected address. Requires customer role."
      parameters: []
      responses: {  }
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the addresses table.'
                  example: architecto
              required:
                - address_id
...